Expand ContentSanitizable to encompass annotations
