+++ app/models/text_block.rb
@@ -1,5 +1,5 @@
 class TextBlock < ApplicationRecord
-  include ContentSanitizable
+  include ContentAnnotatable
 
   # NOTE: This absolutely must be called before all the includes below. If you
   #   put it below them, you will get an ActiveRecord::RecordNotDestroyed
